Fix chat id type in db model.

+++ src/chat_scheduler/models.py
@@ -4,7 +4,7 @@ from django.utils.translation import ugettext_lazy as _
 
 
 class Chat(models.Model):
-    chat_id = models.IntegerField(_("Chat id"), unique=True)
+    chat_id = models.BigIntegerField(_("Chat id"), unique=True)
     created_at = models.DateTimeField(
         _("Created at"),
         auto_now_add=True,
